Update seed interaction in Map
Added comments to the current add and remove Seed functions, and added more conditionals
to the remove Seed function. Since the Java ArrayList remove() function returns a boolean, not the removed object,
I refactored the remove Seed function to use get() to return the Seed, then remove it from seedList with remove().