Initial letter reading implementation
Created a letter class that will display the current letter based on game progress.
I was able to get the conditional that displays the letter to execute only code only once a map is completed,
but there is still an issue with getting the letter to display on the screen.